util: testing: Moved `tests/test_docstrings.py` into `dffml/util/testing/docstrings.py`
